+++ src/views/templates/admin.php
@@ -38,7 +38,7 @@
 		<thead>
 			<tr>
 				<!-- ko foreach: columns -->
-					<th data-bind="css: {sortable: sortable,
+					<th data-bind="visible: visible, css: {sortable: sortable,
 	'sorted-asc': (column_name == $root.sortOptions.field() || sort_field == $root.sortOptions.field()) && $root.sortOptions.direction() === 'asc',
 	'sorted-desc': (column_name == $root.sortOptions.field() || sort_field == $root.sortOptions.field()) && $root.sortOptions.direction() === 'desc'}">
 						<!-- ko if: sortable -->
@@ -58,7 +58,7 @@
 							css: {result: true, even: $index() % 2 == 1, odd: $index() % 2 != 1,
 									selected: $data[$root.primaryKey].raw == $root.itemLoadingId()}">
 					<!-- ko foreach: $root.columns -->
-						<td data-bind="html: $parentContext.$data[column_name].rendered"></td>
+						<td data-bind="html: $parentContext.$data[column_name].rendered, visible: visible"></td>
 					<!-- /ko -->
 				</tr>
 			<!-- /ko -->