+++ changelog.md
@@ -1,5 +1,6 @@
 ## Changelog
 
+- Visible option for columns that accepts either a boolean or closure
 - Relationship constraints now work with hasMany and hasOne fields
 - There is now an `options_filter` option for relationship fields that lets you modify the query before getting the relationship options
 - Custom actions and saves now rebuild the supplied config file after performing the action